Clouseau: Everyone welcome clouseau, a tool to debug UI applications.
It's inspired (and currently not very different) by Android's
hierarchy viewer but it'll hopefully grow to be more than that.
It currently lets you walk the evas object tree, query some evas object
parameters and highlight the object on the canvas.

It currently does not take clipping into account, so you only see the
actual object size, and you can't see the clip frame just yet.

Read the README for a quick how to.

It's very basic atm, read the TODO if you care about future plans, this
is more of a PoC with some useful pieces of code than anything else, but
still, it's more than usable (I wrote it cause I needed it).

Feel free to contribute/sumbit suggestions.

SVN revision: 62740