Clouseau: Fix for class layers displaying.
The bugs happens when a class layer doesn't contain lists (like edje).
The item will be set as tree but the text_get function will check if it
can be compacted and will return true. So we will have the info inside
edje displayed twice, on the layer line and as items of the layer.

The fix forces the layer item to be a tree, no matter the nature of the
contents. It has been done by assigning an item class different for
layers and for infos inside layers. The text_get of the layer item class
just returns the name of the layer without compaction.