Improving Eolian param handling
Instead of receiving class/func/param names as char * we now have
their Eolian type. That way we can easily get info as type and name of
params.
We now also have new type - CLOUSEAU_OBJ_RET. For Eolian funcs that dont
define "pram type" for their return value we dont have part of the info
that we have on normal Eolian param (such as name). For that case we use OBJ_RET.