Refactor GUI for profiles
Before, profiles window was created first and, when the profile was
selected, the window was deleted and the main window was created. The
problem was that the second window was created behind all the existing
windows of the desktop.
Now, the main and profile windows are generated. The profile one is set
as modal, so the main is not accessible.