Introduce focus graph checker
This module checks that all managers which are known to the canvas have
a probebly setted up graph. Which means that all widgets of a graph are
fully connected(no relations to the outside of that graph, and no
seperated widgets which are not accessable via relations). It also
checks for the case of multiple manager objects, that those manager
objects are connected via redirects, and those connections & managers
are representing a tree.

ERR messages are counted as failures found by the checker, CRIT messages
are assertions from the checker which are not met. INF messages are
informations (mostly that a manager is checked etc.).

How the checks are performed is written as comment above the particular
checks.