clouseau_client: change the name of those symbols,
because of the preload, and the fact that _add is also used in
clouseau_app.c it was not possible to connect to a deamon while beeing in
a clouseau_start environment.

This enables the usecase of starting enlightenment with "clouseau_start
enlightenment_start" now you can connect with clouseau_client to your e
and debug the widgets.