Specify Clouseau library to link the extension
During compilation of the objects introspection extension, the
clouseau_debug library is needed.
In Ubuntu machines, it seems that this MUST be specified in the
CMakeLists.txt. Without it, dlopen the extension fails on trying to find
a function present in the Clouseau library.

Thanks to Woohyun for using Ubuntu :-)

@fix